requirement.txt
pillow
keras
requests
tensorflow
keras
numpy
matplotlib
